+++ CHANGELOG.md
@@ -12,5 +12,6 @@ and this project adheres to
 
 - ✨(tools) extract domain from email address #2
 - ✨(oidc) add the authentication backends #2
+- ✨(oidc) add refresh token tools #3
 
 [unreleased]: https://github.com/suitenumerique/django-lasuite/commits/main/